+++ ChangeLog.md
@@ -3,6 +3,7 @@
 - Version 1.0 is explained in the SIGGRAPH Asia 2021 paper (see README.md).
 - [Matlab] Updated calibration: added parameter on cortical magnification and 
   optimized for the weight of the transient channel
+- [Matlab] Added support for reading videos from files (as the python version)
 - [Matlab] Metric now reverts to CPU of no CUDA CPU available
 - [Matlab] added "foveated" parameter to 'fvvdp' (instead of an option)
 - [Matlab] raised ambient light levels for most displays to 250 lux (recommended office illumination)